pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "git-wrapper"
version = "0.1.0"
description = "Minimal git wrapper that injects GitHub App installation tokens."
readme = "Readme.md"
requires-python = ">=3.9"
dependencies = [
    "PyJWT>=2.8.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
cli = [
    "click>=8.1.7",
]

[project.scripts]
git-wrapper = "git_wrapper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/git_wrapper"]

[tool.hatch.build.targets.sdist]
include = [
    "src/git_wrapper",
    "Readme.md",
    "pyproject.toml",
]
